# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Vector API
  description: |
    Vector 지식 관리 플랫폼 API. 파일 업로드/다운로드, 이벤트 동기화,
    충돌 관리, 첨부파일 처리를 지원합니다.
  version: "1.0.0"
  contact:
    name: Vector
servers:
  - url: "http://localhost:3000"
    description: Local development
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, statusCode]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: name is required
            statusCode:
              type: integer
              example: 400

    # ─── 볼트 ────────────────────────────────
...
    UploadResult:
      type: object
      required: [id, path, hash, size_bytes, version]
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        hash:
          type: string
          description: SHA-256 해시
        size_bytes:
          type: integer
          nullable: true
        version:
          type: integer
        auto_merged:
          type: boolean
          description: 3-way merge 자동 병합 여부
        merge_type:
          type: string
          enum: [auto, normal]
          description: 병합 타입 (auto = 자동 병합)
    FileInfo:
      type: object
      required: [id, path, hash, size_bytes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        hash:
          type: string
        size_bytes:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FileDetail:
      type: object
      required: [id, path, hash, size_bytes, content, version]
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
        hash:
          type: string
        size_bytes:
          type: integer
          nullable: true
        content:
          type: string
          nullable: true
        file_type:
          type: string
          enum: [markdown, attachment]
        version:
          type: integer
    FileUploadRequest:
      type: object
      required: [path, hash]
      properties:
        path:
          type: string
        content:
          type: string
        hash:
          type: string
          description: SHA-256 해시
        base_hash:
          type: string
          description: 충돌 감지용 클라이언트 해시
    DeleteResponse:
      type: object
      required: [message, deleted, path]
      properties:
        message:
          type: string
          example: File deleted
        deleted:
          type: boolean
        path:
          type: string
...
    BaseContentResponse:
      type: object
      required: [content, hash]
      properties:
        content:
          type: string
        hash:
          type: string

    # ─── 검색 ────────────────────────────────
...
    DiffOperation:
      type: object
      required: [op, text]
      properties:
        op:
          type: integer
          description: "0=equal, 1=insert, -1=delete"
        text:
          type: string
    ConflictResult:
      type: object
      required: [conflict, current_hash, incoming_hash, conflict_path]
      properties:
        conflict:
          type: boolean
          enum: [true]
        current_hash:
          type: string
        incoming_hash:
          type: string
        conflict_path:
          type: string
        base_hash:
          type: string
        diff:
          type: array
          items:
            $ref: "#/components/schemas/DiffOperation"
        can_auto_merge:
          type: boolean
...
  parameters:
    VaultId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 볼트 UUID
    FilePath:
      name: filePath
      in: path
      required: true
      schema:
        type: string
      description: 파일 경로 (와일드카드)
paths:
...
  /v1/vault/{id}/file:
    put:
      operationId: uploadFile
      summary: JSON 메타데이터로 파일 업로드
      tags: [Files]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - name: x-device-id
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUploadRequest"
      responses:
        "200":
          description: 업로드 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "409":
          description: 충돌 감지
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictResult"
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/vault/{id}/file/{filePath}:
    get:
      operationId: getFile
      summary: 파일 내용 조회
      tags: [Files]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - $ref: "#/components/parameters/FilePath"
        - name: base
          in: query
          schema:
            type: string
          description: 특정 해시의 버전 내용 조회
      responses:
        "200":
          description: 파일 상세 정보
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/FileDetail"
                  - $ref: "#/components/schemas/BaseContentResponse"
        "404":
          description: 파일 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

    delete:
      operationId: deleteFile
      summary: 파일 소프트 삭제
      tags: [Files]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - $ref: "#/components/parameters/FilePath"
        - name: x-device-id
          in: header
          schema:
            type: string
      responses:
        "200":
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: 파일 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/vault/{id}/files:
    get:
      operationId: listFiles
      summary: 파일 목록 조회
      tags: [Files]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 파일 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileInfo"
...
  /v1/vault/{id}/attachment/{filePath}:
    put:
      operationId: uploadAttachment
      summary: 바이너리 첨부파일 업로드 (MinIO)
      tags: [Attachments]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - $ref: "#/components/parameters/FilePath"
        - name: x-device-id
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/gif:
            schema:
              type: string
              format: binary
          image/svg+xml:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
          application/pdf:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: 업로드 성공 (id, path, version 반환)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  path:
                    type: string
                  version:
                    type: integer
        "413":
          description: 본문 크기 초과
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

    get:
      operationId: getAttachment
      summary: 바이너리 첨부파일 조회
      tags: [Attachments]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/VaultId"
        - $ref: "#/components/parameters/FilePath"
      responses:
        "200":
          description: 바이너리 파일 내용
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: 파일 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"

  # ─── 충돌 관리 ────────────────────────────
...